[2019] Add day 9.

+++ 2019/src/advent/intcode/Result.scala
@@ -2,7 +2,7 @@ package advent.intcode
 
 sealed trait Result
 
-case class Success(value: Int) extends Result
+case class Success(value: Long) extends Result
 
 case class Failure(message: String) extends Result
